ai-trading-system/requirements.txt
annotated-types==0.7.0
certifi==2025.10.5
charset-normalizer==3.4.3
colorama==0.4.6
duckdb==1.4.1
greenlet==3.2.4
idna==3.11
iniconfig==2.1.0
markdown-it-py==4.0.0
mdurl==0.1.2
numpy==2.3.3
packaging==25.0
pandas==2.3.3
pluggy==1.6.0
pydantic==2.12.0
pydantic_core==2.41.1
Pygments==2.19.2
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
rich==14.2.0
six==1.17.0
SQLAlchemy==2.0.44
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0

jsonschema==4.23.0
ollama==0.6.0
httpx==0.28.1
anyio==4.11.0
httpcore==1.0.9
h11==0.16.0
sniffio==1.3.1

